Refactor tx modules/methods
