+++ ocular/src/chain/client.rs
@@ -11,7 +11,6 @@ use tendermint_rpc::{self, WebSocketClient, WebSocketClientDriver};
 use super::ChainName;
 
 pub mod airdrop;
-pub mod authz;
 pub mod cache;
 pub mod grpc;
 pub mod query;
